Move move formatter to separate file
